Attemp to resolved CID 143853 Argument cannot be negative (NEGATIVE_RETURNS) fread requires size_t for the buffer size. a positive long can be stores into size_t.

+++ src/shared/Config/Config.cpp
@@ -27,7 +27,8 @@ bool ConfigFile::openAndLoadConfigFile(std::string configFileName)
 
         // go to the end of the file
         fseek(configFile, 0, SEEK_END);
-        uint32_t fileLength = ftell(configFile);
+
+        long fileLength = ftell(configFile);
         if (fileLength <= 0)
         {
             fclose(configFile);
